Implement Billing component - Add stripe integration to billing page - restructure component folder
