+++ frontend/src/components/Sidebar/Sidebar.styling.tsx
