+++ frontend/src/components/__tests__/Sidebar.spec.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Sidebar from '../shared_components/Sidebar/Sidebar';
+import Sidebar from '../Sidebar/Sidebar';
 import { cleanup } from 'react-testing-library';
 import { renderWithRouter } from '../../helpers/functions';
 import 'jest';
