--- frontend/src/pages/Billing.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-const Billing = () => {
-  return (
-    <>
-    </>
-  );
-};
-
-export default Billing;
